+++ digid_eherkenning/oidc/admin.py
@@ -2,16 +2,19 @@ from collections.abc import Sequence
 from copy import deepcopy
 
 from django.contrib import admin
+from django.forms import modelform_factory
 from django.utils.translation import gettext_lazy as _
 
+from mozilla_django_oidc_db.constants import OIDC_MAPPING
+from mozilla_django_oidc_db.forms import OpenIDConnectConfigForm
 from solo.admin import SingletonModelAdmin
 
-from .forms import admin_modelform_factory
 from .models import (
     DigiDConfig,
     DigiDMachtigenConfig,
     EHerkenningBewindvoeringConfig,
     EHerkenningConfig,
+    OpenIDConnectBaseConfig,
 )
 
 # Using a dict because these retain ordering, and it makes things a bit more readable.
@@ -58,6 +61,36 @@ COMMON_FIELDSETS = {
         "classes": ["collapse in"],
     },
 }
+
+
+def admin_modelform_factory(model: type[OpenIDConnectBaseConfig], *args, **kwargs):
+    """
+    Factory function to generate a model form class for a given configuration model.
+
+    The configuration model is expected to be a subclass of
+    :class:`~digid_eherkenning_oidc_generics.models.OpenIDConnectBaseConfig`.
+
+    Additional args and kwargs are forwarded to django's
+    :func:`django.forms.modelform_factory`.
+    """
+    kwargs.setdefault("form", OpenIDConnectConfigForm)
+    Form = modelform_factory(model, *args, **kwargs)
+
+    assert issubclass(
+        Form, OpenIDConnectConfigForm
+    ), "The base form class must be a subclass of OpenIDConnectConfigForm."
+
+    # update the mapping of discovery endpoint keys to model fields, since our base
+    # model adds the ``oidc_op_logout_endpoint`` field.
+    Form.oidc_mapping = {
+        **deepcopy(OIDC_MAPPING),
+        "oidc_op_logout_endpoint": "end_session_endpoint",
+    }
+    Form.required_endpoints = [
+        *Form.required_endpoints,
+        "oidc_op_logout_endpoint",
+    ]
+    return Form
 
 
 def fieldsets_factory(claim_mapping_fields: Sequence[str]):
